export mock data & split components
