+++ src/interfaces/business_monitor.ts
@@ -138,4 +138,130 @@ export const mockReports: IMockReport[] = [
     assuranceStandards: "無",
     isVerifiedByThirdParty: true,
   },
+  {
+    id: 8,
+    company: "遊戲橘子數位科技股份有限公司",
+    title: "2024 年永續報告書",
+    reportYear: "2024",
+    period: "2024/01/01 ~ 2024/12/31",
+    industry: "數位雲端",
+    capital: "10億~50億",
+    verificationAgency: "勤業眾信聯合會計師事務所",
+    verificationStandards: "參考國際永續標準、準則與規範(GRI、SASB)",
+    assuranceAgency: "勤業眾信聯合會計師事務所",
+    assuranceStandards: "確信準則 3000 號",
+    isVerifiedByThirdParty: true,
+  },
+  {
+    id: 9,
+    company: "遊戲橘子數位科技股份有限公司",
+    title: "2023 年永續報告書",
+    reportYear: "2023",
+    period: "2023/01/01 ~ 2023/12/31",
+    industry: "數位雲端",
+    capital: "10億~50億",
+    verificationAgency: "勤業眾信聯合會計師事務所",
+    verificationStandards: "參考國際永續標準、準則與規範(GRI、SASB)",
+    assuranceAgency: "勤業眾信聯合會計師事務所",
+    assuranceStandards: "確信準則 3000 號",
+    isVerifiedByThirdParty: true,
+  },
+  {
+    id: 10,
+    company: "遊戲橘子數位科技股份有限公司",
+    title: "2022 年永續報告書",
+    reportYear: "2022",
+    period: "2022/01/01 ~ 2022/12/31",
+    industry: "數位雲端",
+    capital: "10億~50億",
+    verificationAgency: "勤業眾信聯合會計師事務所",
+    verificationStandards: "參考國際永續標準、準則與規範(GRI、SASB)",
+    assuranceAgency: "勤業眾信聯合會計師事務所",
+    assuranceStandards: "確信準則 3000 號",
+    isVerifiedByThirdParty: true,
+  },
+  {
+    id: 11,
+    company: "智冠科技股份有限公司",
+    title: "2024 年永續報告書",
+    reportYear: "2024",
+    period: "2024/01/01 ~ 2024/12/31",
+    industry: "數位雲端",
+    capital: "10億~50億",
+    verificationAgency: "安侯建業聯合會計師事務所",
+    verificationStandards: "參考國際永續標準、準則與規範(GRI、SASB)",
+    assuranceAgency: "安侯建業聯合會計師事務所",
+    assuranceStandards: "確信準則 3000 號",
+    isVerifiedByThirdParty: true,
+  },
+  {
+    id: 12,
+    company: "華義國際數位娛樂股份有限公司",
+    title: "2024 年永續報告書",
+    reportYear: "2024",
+    period: "2024/01/01 ~ 2024/12/31",
+    industry: "數位雲端",
+    capital: "10億以下",
+    verificationAgency: "無",
+    verificationStandards: "參考國際永續標準、準則與規範(GRI)",
+    assuranceAgency: "無",
+    assuranceStandards: "無",
+    isVerifiedByThirdParty: false,
+  },
+  {
+    id: 13,
+    company: "大統資訊科技股份有限公司",
+    title: "2024 年永續報告書",
+    reportYear: "2024",
+    period: "2024/01/01 ~ 2024/12/31",
+    industry: "數位雲端",
+    capital: "10億以下",
+    verificationAgency: "台灣檢驗科技股份有限公司(SGS)",
+    verificationStandards: "參考國際永續標準、準則與規範(GRI、SASB)",
+    assuranceAgency: "資誠聯合會計師事務所",
+    assuranceStandards: "確信準則 3000 號",
+    isVerifiedByThirdParty: true,
+  },
+  {
+    id: 14,
+    company: "鑫傳國際多媒體科技股份有限公司",
+    title: "2024 年永續報告書",
+    reportYear: "2024",
+    period: "2024/01/01 ~ 2024/12/31",
+    industry: "數位雲端",
+    capital: "10億以下",
+    verificationAgency: "無",
+    verificationStandards: "參考國際永續標準、準則與規範(GRI)",
+    assuranceAgency: "無",
+    assuranceStandards: "無",
+    isVerifiedByThirdParty: false,
+  },
+  {
+    id: 15,
+    company: "鈊象電子股份有限公司",
+    title: "2024 年永續報告書",
+    reportYear: "2024",
+    period: "2024/01/01 ~ 2024/12/31",
+    industry: "數位雲端",
+    capital: "10億~50億",
+    verificationAgency: "勤業眾信聯合會計師事務所",
+    verificationStandards: "參考國際永續標準、準則與規範(GRI、SASB)",
+    assuranceAgency: "勤業眾信聯合會計師事務所",
+    assuranceStandards: "確信準則 3000 號",
+    isVerifiedByThirdParty: true,
+  },
+  {
+    id: 16,
+    company: "樂意傳播股份有限公司",
+    title: "2024 年永續報告書",
+    reportYear: "2024",
+    period: "2024/01/01 ~ 2024/12/31",
+    industry: "數位雲端",
+    capital: "10億以下",
+    verificationAgency: "無",
+    verificationStandards: "參考國際永續標準、準則與規範(GRI)",
+    assuranceAgency: "無",
+    assuranceStandards: "無",
+    isVerifiedByThirdParty: false,
+  },
 ];
